systemc: Stop including mem/port.hh from the systemc headers.
The systemc headers are supposed to be hermetic, so that they can be
included from generic systemc code without any dependency on the gem5
code base, at least when the systemc components are compiled, outside
of those header files themselves. It should be possible to copy the
"ext" directory out of the tree and make it available to systemc code,
and that code should be compileable.

Change-Id: Iec16a534ac04c7895cd8a30940b0acf64c257dde
Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/5995
Reviewed-by: Earl Ou <shunhsingou@google.com>
Tested-by: Earl Ou <shunhsingou@google.com>
(cherry picked from commit 45c469fce538934e97692c0e1b06f5d4c14078bd)
Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/6335