scons: Use either the "build" or "gem5.build" as build anchor.
If gem5.build already exists, then that build directory can be used
without having to worry about variants.

If it doesn't exist and we find a build/${VARIANT} style path, then we
use that as the anchor.

In either case, the variant name is the final component of the build
path. The parse_build_path function had been separating that out, but it
was just put back onto the path again anyway by the only caller, and
then split out again when that path was consumed. We save a step by not
splitting it out in parse_build_path.

Change-Id: I8705b3dbb7664748f5525869cb188df70319d403
Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/7340
Reviewed-by: Yu-hsin Wang <yuhsingw@google.com>
Presubmit-Ready: Gabe Black <gabeblack@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>