mem: add extension mechanism into Packet
Create the extension fields to carry additional user-defined information.
Each type of the extension will have a unique extension ID and there is
a linked list of extension in every packet. There will be most one
extension with the same type in the linked list. With the RefCountingPtr,
the extension will be deleted automatically. That is, the caller should
allocate the extension and add into the packet.

Change-Id: I84b04e88dbf696387292a93352b2fd6953c5a6ec
Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/7372
Reviewed-by: Yu-hsin Wang <yuhsingw@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Presubmit-Ready: Yan Lee <yanlee@google.com>