scons: Flesh out the help text for "magic" targets.
These targets are not necessarily obvious, and tell SCons to do useful
things, like build a particular version of the gem5 binary with a
particular configuration, or run the unit tests.

Add descriptions of these targets to the help so that they are much
more discoverable.

Change-Id: If84399be1a7155ff5f66f511efe1f1c241089c84
Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/7313
Reviewed-by: Yu-hsin Wang <yuhsingw@google.com>
Presubmit-Ready: Gabe Black <gabeblack@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>