systemc: modify get_dir_mem_ptr behavior
According to the spec, it seems like we shouldn't touch the
data_ptr from the trans payload. But in the current
implementation, we send a pkt which would use the data_ptr
from the trans payload to gem5. Therefore, we decide to
allocate another data_ptr buffer for DMI.

Bug: 238684588
Test: Run gem5 with tlm_bridge
Change-Id: Ide09c38ca819fecf8b430e8ec7760b6a13775df3
Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/8472
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jui-min Lee <fcrh@google.com>
Presubmit-Ready: Chang-Yi Kuo <changyikuo@google.com>